+++ src/reanimated2/layoutReanimation/web/index.ts
@@ -1,4 +1,5 @@
 'use strict';
+import type { ReduceMotion } from '../../commonTypes';
 import {
   BounceIn,
   BounceInData,
@@ -36,6 +37,37 @@ import {
 import { ZoomIn, ZoomInData, ZoomOut, ZoomOutData } from './animation/Zoom.web';
 
 import type { AnimationData } from './animationParser';
+
+// Since we cannot remove keyframe from DOM by its name, we have to store its id
+export const customAnimations = new Map<string, number>();
+
+export const WEB_ANIMATIONS_ID = 'ReanimatedWebAnimationsStyle';
+
+export type AnimationCallback = ((finished: boolean) => void) | null;
+
+export interface AnimationConfig {
+  animationName: string;
+  duration: number;
+  delay: number;
+  easing: string;
+  reduceMotion: boolean;
+  callback: AnimationCallback;
+}
+
+export interface CustomConfig {
+  easingV?: () => number;
+  durationV?: number;
+  delayV?: number;
+  randomizeDelay?: boolean;
+  reduceMotionV?: ReduceMotion;
+  callbackV?: AnimationCallback;
+}
+
+export enum TransitionType {
+  LINEAR,
+  SEQUENCED,
+  FADING,
+}
 
 export const AnimationsData: Record<string, AnimationData> = {
   ...FadeInData,
